# Conflicts:
#   src/app/auth.service.ts
#   src/app/contacts/contact-detail/contact-detail.component.html
#   src/app/contacts/contact.service.ts

+++ src/app/contacts/contact-list/contact-list.component.ts
@@ -9,19 +9,10 @@ import {ContactService} from '../contact.service';
 })
 export class ContactListComponent implements OnInit {
   contacts: Contact[];
-  filteredContacts: Contact[];
 
   private contact: Contact;
   private searchTerm: string;
 
-  getSearchTerm() {
-    return this.searchTerm;
-  }
-
-  setSearchTerm(value: string) {
-    this.searchTerm = value;
-    this.filteredContacts = this.filterContacts(value);
-  }
 
   constructor(private contactService: ContactService) {
   }
@@ -29,18 +20,10 @@ export class ContactListComponent implements OnInit {
   ngOnInit() {
     this.contactService.getReq()
       .subscribe((contact: Contact[]) => this.contacts = contact);
-    this.contactService.getReq().subscribe(result => console.log(result));
 
-    this.contactService.deleteContact(this.contact.contactID).subscribe();
-    this.contactService.addContact(this.contact).subscribe(contact => this.contacts.push(contact));
-    this.filteredContacts = this.contacts;
-
+    if (this.contact) {
+      this.contactService.deleteContact(this.contact.contactID).subscribe();
+      this.contactService.addContact(this.contact).subscribe(contact => this.contacts.push(contact));
+    }
   }
-
-  private filterContacts(searchString: string) {
-    return this.contacts.filter(contact =>
-      contact.firstName.toLowerCase().indexOf(searchString));
-  }
-
-
 }
